Improve Agents preview page load when there are no agents
